Move utility to code that uses it.
