+++ opencog/atomutils/CMakeLists.txt
@@ -14,11 +14,11 @@ TARGET_LINK_LIBRARIES(atomutils
 )
 
 IF (CYGWIN)
-    INSTALL (TARGETS atomutils EXPORT AtomSpaceTargets
+	INSTALL (TARGETS atomutils EXPORT AtomSpaceTargets
 		DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
 	)
 ELSE (CYGWIN)
-    INSTALL (TARGETS atomutils EXPORT AtomSpaceTargets
+	INSTALL (TARGETS atomutils EXPORT AtomSpaceTargets
 		DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
 	)
 ENDIF (CYGWIN)
@@ -31,6 +31,5 @@ INSTALL (FILES
 	FuzzyMatchBasic.h
 	ThreadSafeHandleMap.h
 	Neighbors.h
-	Substitutor.h
 	DESTINATION "include/opencog/atomutils"
 )
