+++ opencog/query/Substitutor.h
